bugfix: add item bugfix: add file item bugfix: allow professor instead of responsible for execution course bugfix: add missing methods to Teachar Administration action

+++ src/net/sourceforge/fenixedu/domain/functionalities/AvailabilityPolicy.java
@@ -51,7 +51,9 @@ public abstract class AvailabilityPolicy extends AvailabilityPolicy_Base {
             super.afterRemove(availabilityPolicy, accessibleItem);
 
             if (accessibleItem != null) {
-                availabilityPolicy.delete();
+                if (availabilityPolicy != null) {
+                    availabilityPolicy.delete();
+                }
             }
         }
